requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.1
cow==1.0.7
cowsay==6.1
cryptography==45.0.6
cycler==0.12.1
decorator==5.2.1
distlib==0.3.9
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-filter==25.1
django-rest-authtoken==2.1.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
Faker==37.0.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
instaloader==4.14.1
kiwisolver==1.4.8
magic-filter==1.0.12
matplotlib==3.10.1
moviepy==2.2.1
multidict==6.2.0
numpy==2.2.4
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.6
prettytable==3.16.0
proglog==0.1.12
propcache==0.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.29.1
wcwidth==0.2.13
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.18.3